Core in progress

+++ Assets/_Client/Source/Features/Letter/Letter.cs
@@ -7,11 +7,11 @@ namespace WordMaster
     {
         public char Value { get; }
         public ReactiveProperty<Vector2> Position { get; }
-        public int Radius { get; }
+        public float Radius { get; }
         public CompositeDisposable Disposables { get; }
         
         public Letter(char value, Vector2 position, 
-            int radius, CompositeDisposable disposables)
+            float radius, CompositeDisposable disposables)
         {
             Value = value;
             Position = new ReactiveProperty<Vector2>(position);
